+++ src/pages/trip-details/create-activity-modal.tsx
@@ -1,13 +1,31 @@
 import { Calendar, Tag, X } from "lucide-react"
 import { Button } from "../../components/button"
 import { FormEvent } from "react"
+import { api } from "../../lib/axios"
+import { useParams } from "react-router-dom"
 
 interface CreateActivityModalProps{
     closeCreateActivityModal: () => void
-    createActivity: (event: FormEvent<HTMLFormElement>) => void
 }
 
-export function CreateActivityModal({closeCreateActivityModal, createActivity}: CreateActivityModalProps) {
+export function CreateActivityModal({closeCreateActivityModal}: CreateActivityModalProps) {
+    const { tripId } = useParams()
+    
+    async function createActivity(event: FormEvent<HTMLFormElement>) {
+        event.preventDefault()
+
+        const data = new FormData(event.currentTarget)
+        const title = data.get('title')?.toString()
+        const occurs_at = data.get('occurs_at')?.toString()
+        
+        await api.post(`/trips/${tripId}/activities`, {
+            title,
+            occurs_at
+        })
+
+        closeCreateActivityModal()
+    }
+
     return (
         <div className="fixed inset-0 bg-black/60 flex items-center justify-center">
             <div className="bg-zinc-900 w-[640px] py-5 px-6 shadow-shape rounded-xl space-y-5">
