--- src/types.d.ts
@@ -1,8 +1,0 @@
-interface Activity {
-  date: string;
-  activities: {
-    id: string;
-    title: string;
-    occurs_at: string;
-  }[];
-}
